+++ test_project/generic/api.py
@@ -12,31 +12,20 @@ from .models import ModelA, ModelB
 class ModelASerializer(serializers.ModelSerializer):
     class Meta:
         model = ModelA
-        fields = [
-            'name',
-            'rel_b',
-        ]
+        fields = ["name", "rel_b"]
 
 
 class ModelBSerializer(serializers.ModelSerializer):
     class Meta:
         model = ModelB
-        fields = [
-            'name',
-            'a',
-            'content_type',
-            'object_id',
-        ]
+        fields = ["name", "a", "content_type", "object_id"]
 
 
 class ModelAViewSet(ModelViewSet):
     serializer_class = ModelASerializer
     queryset = ModelA.objects.all()
     filter_backends = [DjangoFilterBackend]
-    filter_fields = [
-        'name',
-        'rel_b',
-    ]
+    filter_fields = ["name", "rel_b"]
 
 
 class ModelBViewSet(ModelViewSet):