+++ test_project/many_to_many/alchemy.py
@@ -9,7 +9,7 @@ from ..alchemy import Base
 
 
 class Publication(Base):
-    __tablename__ = 'many_to_many_publication'
+    __tablename__ = "many_to_many_publication"
     id = Column(Integer, primary_key=True)
     title = Column(String(30))
 
@@ -19,23 +19,23 @@ class Publication(Base):
 
 
 publication_article_association_table = Table(
-    'many_to_many_article_publications',
+    "many_to_many_article_publications",
     Base.metadata,
-    Column('id', Integer),
-    Column('publication_id', Integer, ForeignKey('many_to_many_publication.id')),
-    Column('article_id', Integer, ForeignKey('many_to_many_article.id')),
+    Column("id", Integer),
+    Column("publication_id", Integer, ForeignKey("many_to_many_publication.id")),
+    Column("article_id", Integer, ForeignKey("many_to_many_article.id")),
 )
 
 
 class Article(Base):
-    __tablename__ = 'many_to_many_article'
+    __tablename__ = "many_to_many_article"
     id = Column(Integer, primary_key=True)
     headline = Column(String(100))
 
     publications = relationship(
         Publication,
         secondary=publication_article_association_table,
-        backref=backref('articles', uselist=True),
+        backref=backref("articles", uselist=True),
         uselist=True,
     )