+++ test_project/one_to_one/models.py
@@ -16,8 +16,7 @@ class Place(models.Model):
 
 @six.python_2_unicode_compatible
 class Restaurant(models.Model):
-    place = models.OneToOneField(Place, primary_key=True,
-                                 on_delete=models.CASCADE)
+    place = models.OneToOneField(Place, primary_key=True, on_delete=models.CASCADE)
     serves_hot_dogs = models.BooleanField(default=False)
     serves_pizza = models.BooleanField(default=False)